requirements.txt
absl-py==2.1.0
aioice==0.9.0
aiortc==1.10.1
altair==5.5.0
attrs==25.1.0
av==13.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dnspython==2.7.0
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.6.0
idna==3.10
ifaddr==0.2.0
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
MouseInfo==0.1.3
mpmath==1.3.0
narwhals==1.26.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.6
pyarrow==19.0.0
PyAutoGUI==0.9.54
pycparser==2.22
pydeck==0.9.1
pyee==12.1.1
PyGetWindow==0.0.9
Pygments==2.19.1
pylibsrtp==0.11.0
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rubicon-objc==0.5.0
scipy==1.15.1
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
streamlit==1.42.0
streamlit-webrtc==0.47.9
sympy==1.13.1
tenacity==9.0.0
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
